+++ WVC.Models/Models/wvc_wood_volum_item.cs
@@ -16,19 +16,19 @@ namespace WVC.Models
         public string description { get; set; }
 		
 		
-        public Nullable<float> length { get; set; }
+        public Nullable<decimal> length { get; set; }
 		
 		
-        public Nullable<float> girth { get; set; }
+        public Nullable<decimal> girth { get; set; }
 		
 		
-        public Nullable<float> volume { get; set; }
+        public Nullable<decimal> volume { get; set; }
 		
 		
-        public Nullable<float> co_efficient { get; set; }
+        public Nullable<decimal> co_efficient { get; set; }
 		
 		
-        public Nullable<float> final_volume { get; set; }
+        public Nullable<decimal> final_volume { get; set; }
         public virtual wvc_wood_volume wvc_wood_volume { get; set; }
     }
 }
